pyproject.toml
[project]
name = "ai-agent-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.12.0",
    "ipython>=9.5.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.7",
    "langgraph-api>=0.4.26",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "tiktoken>=0.11.0",
    "wikipedia>=1.4.0",
]
